+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 pnet-data-api 2.7.0
 ===================
 
-Compatibility with Partner.Net 6.44.
+Needs Partner.Net 6.44.
 
-* Add the division label to person advisor assignments.
+* Add advisors to companies.
+* Add the division label to advisor assignments.
 * Fixed a lot of warnings.
 
 pnet-data-api 2.6.0